Added a header param / set_header method
There are bound to be special use cases for using the PHP header()
function (or EE's set_header()) verbatim instead of adding
params/methods for every exception. If one is already using a plugin
for headers they shouldn't have to enable PHP in a template to do so,
so this should be made available (at least for custom use, not
necessarily in the public plugin).